+++ src/runtime/RuntimeHelper.php
@@ -94,12 +94,21 @@ class RuntimeHelper implements RuntimeHelperInterface
     }
 
     /**
-     * @param string $content
+     * @param mixed $content
      * @return string
      */
-    public static function htmlEncode(string $content): string
+    public static function htmlEncode($content): string
     {
-        return CompilerHelper::htmlEncode($content);
+        if (is_string($content)) {
+            return CompilerHelper::htmlEncode($content);
+        }
+        if (is_int($content) || is_bool($content) || is_float($content)) {
+            return (string)$content;
+        }
+        if (is_array($content)) {
+            return '[Array]';
+        }
+        throw new \InvalidArgumentException('Unsupported type');
     }
 
     /**