+++ src/compile/Compiler.php
@@ -31,7 +31,7 @@ class Compiler implements CompilerInterface
 {
     private const A_BUBBLE = Parser::ACTION_BUBBLE_THE_ONLY;
 
-    private const VERSION = '0.2.0-dev.1';
+    private const VERSION = '0.2.0';
 
     private const STRING_ESCAPE_LETTER = [
         'b' => "\x08",